Validacao no campo cadastro, e aglumas alterações no dbhelp

+++ app/src/main/java/dev/expertsunited/mybooks/model/Usuario.java
@@ -1,7 +1,7 @@
 package dev.expertsunited.mybooks.model;
 
 public class Usuario {
-    private Double id;
+    private Integer  id;
     private String nome;
     private String email;
     private String login;
@@ -11,11 +11,9 @@ public class Usuario {
 
     }
 
-    public Double getId() {
-        return id;
-    }
+    public Integer getId() { return id; }
 
-    public void setId(Double id) {
+    public void setId(Integer id) {
         this.id = id;
     }
